projeto02_painel/requirements.txt
flask
psutil
qrcode
pywebview
